refactor(discovery): split pipeline.go (1394 → 4 files) by concern
Sixth worst-first file-size refactor.

  pipeline.go            450  Pipeline + Start/Cancel + run + state helpers
  pipeline_types.go      292  enums + payloads + PipelineConfig tree
  pipeline_defaults.go   302  port presets + DefaultConfig + top-1000 list
  pipeline_phases.go     396  per-phase runners + scanning helpers

All under the 600-line warn budget.

Drive-by cleanup forced by --new-from-rev on relocated lines:
extracted four phaseNameXxx constants (enumeration / resolution /
scanning / assessment) and threaded them through every phase payload,
PhaseDurations map key, and runner call. Silences goconst (the
literals appear 4-6 times across the moved lines).

Build, vet, lint (0 new issues), discovery tests all pass.
